SW-5896: Update Glossary to Use Sphinx Glossary Feature
This patch updates the glossary to use the Sphinx built-in glossary
feature.

The existing glossary uses manually created labels and labeled
cross-references of the form :ref:`well used term
<glossary-well-used-term>`, which is cumbersome and throws a 'malformed
table' error when the glossary term is used in a table.

glossary.rst was formatted with the reST definition-list markup and
cross-references throughout the document were changed to the more
compact :term:`well used term` form used by the Sphinx glossary feature.

Signed-off-by: Rob Johnson <gtgprods@gmail.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>