SW-5862: greybus-spec: reserve bundle id 0xff
Each bundle defined in a manfiest has a unique one-byte Bundle ID.
Bundle IDs start with 0x00 and are numbered consecutively.  At this
time there is no reason to believe we will ever get close to 256
distinct bundles in a single interface manifest.

Set aside bundle ID value 0xff as reserved and invalid.  I found a
need for such an invalid value when working on SW-5335, and none was
defined.

Signed-off-by: Alex Elder <elder@linaro.org>
Reviewed-by: Viresh Kumar <viresh.kumar@linaro.org>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>