SW-4385: Add MSC for Control Connection Closure for Suspend Sequence
Add a Message Sequence Chart for the Control Connection Closure for
Suspend Sequence to help describe the Interface Lifecyle and the
interaction between various components in the system.

This patch is one of a series of independent commits intended to make it
easier to accept or reject specific MSCs.

This patch was moved from JIRA SW-4379 to SW-4385.

Signed-off-by: Rob Johnson <gtgprods@gmail.com>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>