SW-7131: camera: split tables in side-by-side layout
Some tables (table-camera-capabilities-tags, table-camera-capture-tags,
and table-camera-metadata-tags) have to many rows to fix in a single
page in the latexpdf target. That make then impossible to be used as
reference.

With the use of adjustbox package to handle tables scaling and with
splitting the tables in a side-by-side layout this big tables fit in a
single page making the content visible.

Signed-off-by: Rui Miguel Silva <rui.silva@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>