SW-7284: firmware: Add 'status' byte to backend version response
The backend device on an Interface may not be ready by the time AP tries
to get its firmware version. The specification already had some support
for that, but by overloading the operation status byte, which we aren't
doing anymore.

Define a new 1-byte value 'status' in the response, which can be used by
the Interface firmware to return a bunch of errors.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Reviewed-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>