SW-7284: firmware: Add new status types for backend updated request
This adds two new status types for the Interface backend firmware updated
request.

STATUS_RETRY will be used if the Interface and the backend device aren't
ready for the update yet and want the AP to try the operation again.

STATUS_NOT_SUPPORTED will be used if the "firmware-tag" doesn't match
any of the required backend firmware packages.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Reviewed-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>