SW-6891: bundle: ignore setting an already active state
Due to multiple requests and negative impact this has on the degree
of kernel code complication, requesting a power state transition for
a bundle which already is in the requested state shall no longer be
treated as an error condition.

Currently the error code returned in this situation is the same as for
requests sent when the bundle is in an invalid state in relation to
the requested state (e.g.: suspend requested for a deactivated
bundle). The latter is an obvious error condition, while the former
is not.

This has no negative impact, as such a request shall be silently
ignored by the interface.

Modify the bundle power management operations to reflect that.

Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>