SW-6939: bootrom: the Ping Operation is actually unimplemented
This patch fixes a bug in the Bootrom Protocol definition which is
needed to support a new Connection closure sequence.

The boot ROM in the Gen 1 ASICs is bootrom version 1b3b767 ("Bootrom:
public key files"), from 16 Oct 2015.

At that time, the Greybus Specification still used 0x00 as the "Invalid"
Operation Type: 0x00 wasn't permitted as a legal Operation Type until
81628c7 ("greybus-spec: change invalid operation type") on 30 March
2016.

Thus, the current listing of the Ping Operation in the Bootrom Protocol
is bogus. Delete it, and fix the language in connection.rst to make it
clear that not all Protocols implement this Operation (note that Camera
Data doesn't either).

Don't bother fixing up the rest of the capitalization in the Ping
definition, as it's about to disappear in a future patch.

Testing Done: built spec without warnings.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Acked-by: Viresh Kumar <viresh.kumar@linaro.org>
Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>