SW-6939: replace "Ping" with "CPort Shutdown"
This is mostly a one-for-one replacement of the Ping Operation with a
new "CPort Shutdown" Operation, with one exception in the SVC Protocol.

The important part of the patch is in source/connection.rst, which
defines this new Operation. Currently, it is completely equivalent to
the existing Ping Operation, except that its Request contains a
"shutdown_phase" field. This field is currently always equal to one, and
has no effect on the recipient, so the only nontrivial effect of this
patch is to make the Request contain a small payload.

The rest of this patch is mostly just the necessary churn in the rest of
the spec to reflect this change, which is a fairly mechanical
search/replace.

The SVC Protocol is an exception: we still need a Ping Operation there
to support an AP/SVC watchdog, which uses Ping as a heartbeat. To
support that, we resurrect the Ping Operation in its old place (Request
Operation Type 0x13), which has the benefit of making spec match
implementation.

Future patches will extend the CPort Shutdown Operation to add a second
phase with different semantics, reflecting its intended role in the new
Connection closure sequences.

Testing Done: build spec without warnings or errors. Grep for "ping" and
              verify that all remaining occurrences are for SVC,
              Loopback, or Timesync, which define their own "ping"
              Operations.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>