SW-2700: bibliography: specify chapter for html
The Bibliography chapter is missing for th html format of the
documentation due to a workaround for latexpdf. However, the doc looks
very awkward and confusing especially for singlehtml - it looks as if
the bibliography is part of the previous chapter. Add html only markup
for the chapter to fix this.

Testing Done:
 - Build & check the output on pdf, html and singlehtml formats

Signed-off-by: David Lin <dtwlin@google.com>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>