SW-7452: camera: Remove Ara name occurrences
Replace "Ara" with "Greybus" in Camera Device Class specifications.

Re-phrase the initial sentences to refer to "Greybus System" as defined
by the Specifications' glossary in place of "Ara system", that was most
probably referring to the Ara phone.
Prefer "Greybus" over "GMP" in this case, because the term is currently
undefined in Specifications.

Signed-off-by: Jacopo Mondi <jacopo.mondi@linaro.org>
Acked-by: Laurent Pinchart <laurent.pinchart@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>