SW-7645: firmware: let module manage sequence of updates
A module may need multiple backend firmware packages for its working and
the AP may not have full knowledge about the order in which those are
required to be updated.

This information is already available with the module bridge firmware
and it shall be allowed to drive this.

Remove firmware-tag from version and backend update requests, so that
the module always update all the backend firmware packages once
requested by the AP.

Reviewed-by: Alex Elder <elder@linaro.org>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>